requirements.txt
torch>=2.0
numpy>=1.24
scipy>=1.10
datasets>=2.14
pyyaml>=6.0
tqdm>=4.65
wandb>=0.15
matplotlib>=3.7
scikit-learn>=1.3
